[quant][refactor tests] Move TestModelNumerics to a separate file Summary: These tests span both QAT and PTQ APIs so factor them out
Test Plan:
python test/test_quantization.py TestModelNumericsEager

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
